* DEL: Removed  messfilter.
