+++ code/engine.vc2008/xrGame/Level.cpp
@@ -170,7 +170,6 @@ CLevel::CLevel():IPureClient	(Device.GetTimerGlobal())
 	m_DemoPlayStoped	= FALSE;
 	m_DemoSave = FALSE;
 	m_DemoSaveStarted = FALSE;
-	m_msg_filter = nullptr;
 	m_demo_info	= nullptr;
 
 	R_ASSERT				(!g_player_hud);
@@ -273,7 +272,7 @@ CLevel::~CLevel()
 			m_reader			= NULL;
 		}
 	}
-	xr_delete(m_msg_filter);
+
 	xr_delete(m_demo_info);
 	if (IsDemoSave())
 	{
