+++ code/engine.vc2008/xrGame/script_action_condition.cpp
@@ -12,3 +12,27 @@
 CScriptActionCondition::~CScriptActionCondition	()
 {
 }
+
+using namespace luabind;
+
+#pragma optimize("s",on)
+void CScriptActionCondition::script_register(lua_State *L)
+{
+	module(L)
+	[
+		class_<CScriptActionCondition>("cond")
+			.enum_("cond")
+			[
+				value("move_end",	int(CScriptActionCondition::MOVEMENT_FLAG)),
+				value("look_end",	int(CScriptActionCondition::WATCH_FLAG)),
+				value("anim_end",	int(CScriptActionCondition::ANIMATION_FLAG)),
+				value("sound_end",	int(CScriptActionCondition::SOUND_FLAG)),
+				value("object_end",	int(CScriptActionCondition::OBJECT_FLAG)),
+				value("time_end",	int(CScriptActionCondition::TIME_FLAG)),
+				value("act_end",	int(CScriptActionCondition::ACT_FLAG))
+			]
+			.def(					constructor<>())
+			.def(					constructor<u32>())
+			.def(					constructor<u32,double>())
+	];
+}
