+++ code/engine.vc2008/xrGame/xrGame.vcxproj.filters
@@ -7382,9 +7382,6 @@
     <ClCompile Include="script_action_condition.cpp">
       <Filter>AI\AScript\ScriptClasses\ScriptActions\script_action_condition</Filter>
     </ClCompile>
-    <ClCompile Include="script_action_condition_script.cpp">
-      <Filter>AI\AScript\ScriptClasses\ScriptActions\script_action_condition</Filter>
-    </ClCompile>
     <ClCompile Include="script_monster_action.cpp">
       <Filter>AI\AScript\ScriptClasses\ScriptActions\script_monster_action</Filter>
     </ClCompile>
